backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.3
aiofiles==23.2.1
httpx==0.26.0
